Move exporter logic to dedicated package
Corresponds to similar changes made in prometheus/memcached_exporter#97.
